+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - added: Added flexible service key matching with URL templating.
+- added: Added support for WebSocket transports for evmRpc plugin.
 - changed: Migrated to JSON-based logs using Pino library.
 - changed: Use serviceKeys for nownodes API key.
 - removed: Removed plugins to reduce load: abstract, amoy, arbitrum, avalanche, base, bobevm, celo, ethereumclassic, ethereumpow, fantom, filecoinfevm, filecoinfevmcalibration, holesky, hyperevm, pulsechain, rsk, sepolia, sonic.